NAS 41 refractor
