+++ src/main/java/uk/gov/crowncommercial/dts/scale/service/agreements/BLL/BusinessLogicClient.java
@@ -423,31 +423,4 @@ public class BusinessLogicClient {
 
         return getOrganisation(orgName);
     }
-
-    public void updateSupplierByDuns(String dunsNumber, SupplierUpdateRequest updateRequest) {
-        supplierService.updateSupplierAndContactByDuns(dunsNumber, updateRequest);
-    }
-
-    public SupplierSummary addSupplierToLotByDuns(String agreementId, String lotId, String dunsNumber) {
-
-        Organisation organisation = supplierService.findOrganisationBySchemeAndEntityId("US-DUNS", dunsNumber);
-        Lot lot = agreementService.findLotByAgreementNumberAndLotNumber(agreementId, lotId);
-        supplierService.addSupplierRelationship(lot, organisation, null, "api", SupplierStatus.ACTIVE);
-        int supplierCount = getLotSuppliers(agreementId, lotId).size();
-        return new SupplierSummary(java.time.LocalDate.now(), "api", supplierCount);
-    }
-
-    public void updateSupplierStatusForLot(String agreementId, String lotId, String dunsNumber, String operation) {
-        Organisation organisation = supplierService.findOrganisationBySchemeAndEntityId("US-DUNS", dunsNumber);
-        Lot lot = agreementService.findLotByAgreementNumberAndLotNumber(agreementId, lotId);
-        SupplierStatus status;
-        if ("suspend".equalsIgnoreCase(operation)) {
-            status = SupplierStatus.SUSPENDED;
-        } else if ("unsuspend".equalsIgnoreCase(operation)) {
-            status = SupplierStatus.ACTIVE;
-        } else {
-            throw new IllegalArgumentException("Invalid operation: " + operation);
-        }
-        supplierService.updateSupplierStatusForLot(lot, organisation, status);
-    }
 }
